requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
scikit-learn
joblib
pillow
pytesseract
pdfplumber
transformers
torch
python-multipart
streamlit
requests
datasets
openai
accelerate
pdf2image
